[Refactor] Refactor ECS folder structure
